SLVS-2009 Refactor sonar dotnet configuration generator logic from connected mode to be reusable by standalone mode
