+++ src/Core/CSharpVB/IRoslynRuleStatus.cs
@@ -18,24 +18,10 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using SonarLint.VisualStudio.Core.CSharpVB;
+namespace SonarLint.VisualStudio.Core.CSharpVB;
 
-// Note: this interface was added as part of the refactoring that was done when
-// the support for configuration of C++ files in Connected Mode was added.
-// It minimised the changes required to the existing binding code that is
-// ruleset-specific, at the cost of downcasts in a couple of places (done by
-// the TryGetRuleSet extension method).
-
-namespace SonarLint.VisualStudio.ConnectedMode.Binding
+public interface IRoslynRuleStatus
 {
-    /// <summary>
-    /// Extends the base binding configuration interface for C#/VB projects where
-    /// the config is expected to have a ruleset
-    /// </summary>
-    public interface ICSharpVBBindingConfig : IBindingConfig
-    {
-        FilePathAndContent<SonarLintConfiguration> AdditionalFile { get; }
-
-        FilePathAndContent<string> GlobalConfig { get; }
-    }
+    string Key { get; }
+    RuleAction GetSeverity();
 }
